Split with_pools.hpp into save and load
