+++ test/extra/persist/test_conversion.cpp
@@ -1,6 +1,6 @@
 #include <catch2/catch_test_macros.hpp>
 
-#include <immer/extra/persist/cereal/with_pools.hpp>
+#include <immer/extra/persist/cereal/save.hpp>
 #include <immer/extra/persist/detail/type_traverse.hpp>
 #include <immer/extra/persist/transform.hpp>
